Beefed up the README.md turned the README.md into pypi docs. Cleaned up the command line to make sessions optional and more flexible. Bumped the version

+++ mongodbrdg/mongodbrdg_main.py
@@ -3,13 +3,16 @@ import argparse
 import pprint
 from datetime import datetime
 import sys
-from mongodbrdg.randomuser import RandomUser, RandomSessions
+from mongodbrdg.randomdata import User, Sessions
 from mongodbrdg.inserter import Inserter
+from mongodbrdg.version import __VERSION__
 import random
 
 
 def main():
-    parser = argparse.ArgumentParser()
+    parser = argparse.ArgumentParser(prog="mongodbrdg",
+                                     description=f"mongodbrdg {__VERSION__}. "
+                                                 f"Generate random JSON data for MongoDB (requires python 3.6).")
     parser.add_argument("--mongodb", default="mongodb://localhost:27017", help="MongoDB host: [default: %(default)s]")
     parser.add_argument("--database", default="USERS", help="MongoDB database name: [default: %(default)s]")
     parser.add_argument("--collection", default="profiles",
@@ -18,34 +21,43 @@ def main():
     parser.add_argument("--batchsize", default=1000, type=int,
                         help="How many docs to insert per batch: [default: %(default)s]")
     parser.add_argument("-locale", default="en", help="Locale to use for data: [default: %(default)s]")
-    parser.add_argument("--seed", type=int, help="Use this seed value to ensure you always get the same data")
+    parser.add_argument("--seed", default=None, type=int, help="Use this seed value to ensure you always get the same data")
     parser.add_argument("--drop", default=False, action="store_true",
                         help="Drop data before creating a new set [default: %(default)s]")
     parser.add_argument("--report", default=False, action="store_true",
                         help="send all generated JSON to the screen [default: %(default)s]")
-    parser.add_argument("--sessions", default=50, type=int,
-                        help="0 to generate a random number of sessions, or a number for a specific number")
+    parser.add_argument("--session", choices=["none", "random", "count"], default="none",
+                        help="Generate a sessions collection [default: %(default)s do not generate] ")
+    parser.add_argument("--sessioncount", default=5, type=int,
+                        help="Default number of sessions to generate."
+                             "Gives the random bound for random sessions [default: %(default)s]")
     parser.add_argument("--sessioncollection", default="sessions",
                         help="Name of sessions collection: [default: %(default)s]")
     parser.add_argument("--bucketsize", type=int, default=1000,
                         help="Bucket size for insert_many [default: %(default)s]")
+    parser.add_argument("--stats", default=False, action="store_true",
+                        help="Report time to insert data")
+
     args = parser.parse_args()
 
     client = pymongo.MongoClient(args.mongodb)
 
     db = client[args.database]
     user_collection = db[args.collection]
-    session_collection = db[args.sessioncollection]
-    session_inserter = Inserter(session_collection, 1000)
     user_inserter = Inserter(user_collection, 1000)
 
-    if args.drop:
-        print(f"Dropping collection: '{user_collection.name}'")
-        db.drop_collection(args.collection)
-        print(f"Dropping collection: '{session_collection.name}'")
-        db.drop_collection(args.sessioncollection)
+    if args.session != "none" :
+        session_collection = db[args.sessioncollection]
+        session_inserter = Inserter(session_collection, 1000)
 
-    user = RandomUser(locale=args.locale, seed=args.seed)
+    if args.drop:
+        print(f"Dropping collection: '{db.name}.{user_collection.name}'")
+        db.drop_collection(args.collection)
+        if args.sessions != "none" :
+            print(f"Dropping collection: '{db.name}.{session_collection.name}'")
+            db.drop_collection(args.sessioncollection)
+
+    user = User(locale=args.locale, seed=args.seed)
 
     try:
         user_doc_count: int = 0
@@ -57,12 +69,20 @@ def main():
             user_inserter.insert(clone)
             if args.report:
                 pprint.pprint(clone)
-            if args.sessions:
-                sessions = RandomSessions(clone["user_id"], clone["registered"])
-                session_count = random.randint(1, args.sessions)
+            if args.session != "none":
+                if args.session == "random":
+                    session_count = random.randint(0, args.sessioncount)
+                else:
+                    session_count = args.sessioncount
+
+                sessions = Sessions(clone["user_id"], clone["registered"])
                 for login, logout in sessions.make_sessions(session_count):
                     session_inserter.insert(login)
+                    if args.report:
+                        pprint.pprint(login)
                     session_inserter.insert(logout)
+                    if args.report:
+                        pprint.pprint(logout)
                     session_doc_count = session_doc_count + 2
                 session_inserter.flush()
 
@@ -74,15 +94,17 @@ def main():
         sys.exit(1)
 
     elapsed = finish - start
-    print("")
     print(f"Inserted {user_doc_count} user docs into {db.name}.{user_collection.name}")
-    print(f"Inserted {session_doc_count} session docs into {db.name}.{session_collection.name}")
-    print(f"Elapsed time: {elapsed}")
-    elapsed_time = float(elapsed.seconds) + float(elapsed.microseconds) / 1000000
-    print(f"Elapsed seconds: {elapsed_time}")
+    if args.session != "none":
+        print(f"Inserted {session_doc_count} session docs into {db.name}.{session_collection.name}")
 
-    docs_per_second = float(user_doc_count + session_doc_count) / elapsed_time
-    print(f"Inserted {round(docs_per_second, 0)} docs per second")
+    if args.stats:
+        print(f"Elapsed time: {elapsed}")
+        elapsed_time = float(elapsed.seconds) + float(elapsed.microseconds) / 1000000
+        print(f"Elapsed seconds: {elapsed_time}")
+
+        docs_per_second = float(user_doc_count + session_doc_count) / elapsed_time
+        print(f"Inserted {round(docs_per_second, 0)} docs per second")
 
 
 if __name__ == "__main__":
